Share assets, i18n and store.
