+++ src/popup/store/index.js
@@ -1,5 +1,3 @@
-/* eslint no-param-reassign:
-["error", { "props": true, "ignorePropertyModificationsFor": ["state"] }] */
 import {
   CURRENT_LIVES,
   ENDED_LIVES,
@@ -9,14 +7,9 @@ import {
   SCHEDULED_LIVES,
   SHOULD_SYNC_SETTINGS,
 } from 'shared/store/keys'
-import Vue from 'vue'
-import Vuex from 'vuex'
-import connect from './connect'
+import createStore from 'shared/store/create'
 
-Vue.use(Vuex)
-
-// Default state will be covered by values of the same key on connecting to background store
-const defaultState = {
+const store = createStore({
   [ENDED_LIVES]: [],
   [CURRENT_LIVES]: [],
   [SCHEDULED_LIVES]: [],
@@ -24,18 +17,6 @@ const defaultState = {
   [LOCALE]: 'en',
   [IS_POPUP_FIRST_RUN]: true,
   [SHOULD_SYNC_SETTINGS]: false,
-}
+})
 
-const store = {
-  state: defaultState,
-  mutations: {
-    ...Object.fromEntries(Object.keys(defaultState).map(
-      key => [key, (state, value) => {
-        state[key] = value
-      }],
-    )),
-  },
-  plugins: [connect],
-}
-
-export default new Vuex.Store(store)
+export default store
