+++ src/shared/.eslintrc.js
@@ -1,0 +1,19 @@
+const path = require('path')
+
+module.exports = {
+  globals: {
+    VERSION: 'readonly',
+  },
+  settings: {
+    'import/resolver': {
+      node: {
+        paths: [__dirname, path.join(__dirname, '..')],
+        extensions: [
+          '.js',
+          '.json',
+          '.vue',
+        ],
+      },
+    },
+  },
+}
