+++ src/shared/store/connect.js
